investigate how to get before edit functionity
my goal is basically to implement formulas.
my goal is two have two values stored in the underlying
data of the table. the first value will be the formula
and the second will be the result.

when i tried passing an actual array value as the value for
an individual cell from my server to hot, my afterChange hook
appeared to stop working entirely. So instead I convert the cell
data to a json string which is then parsed by the beginEditing
function and by my various hooks.

the major change to the beginEditing function is as follows.
the original functionality was that when the function
determines that the user is editing the existing data in the
cell, the variable initialValue is set to this.originalValue.
In my case originalValue is my json string and I just want to
display the first value in it. So I parse the json, and set
initialValue to the first element. The result is that when you
go to edit the cell values, you will see the formula.